+++ gdsfactory/routing/get_bundle_from_steps.py
@@ -104,11 +104,11 @@ def get_bundle_from_steps(
 
     port2 = ports2[0]
     x2, y2 = port2.midpoint
-    orientation = int(port2.orientation)
+    orientation = port2.orientation
 
-    if orientation in [0, 180]:
+    if int(orientation) in [0, 180]:
         waypoints += [(x, y2)]
-    elif orientation in [90, 270]:
+    elif int(orientation) in [90, 270]:
         waypoints += [(x2, y)]
 
     x = cross_section(**kwargs)